+++ st2api/tests/unit/controllers/v1/test_executions_simple.py
@@ -232,7 +232,9 @@ class TestActionExecutionController(FunctionalTest):
         self.assertEqual(post_resp.status_int, 201)
         delete_resp = self._do_delete(self._get_actionexecution_id(post_resp))
         self.assertEqual(delete_resp.status_int, 200)
-        self.assertEqual(delete_resp.json['status'], 'canceled')
+        self.assertEqual(delete_resp.json['status'], 'canceling')
+        expected_result = {'message': 'Action canceled by user.', 'user': 'stanley'}
+        self.assertDictEqual(delete_resp.json['result'], expected_result)
 
     def test_post_delete_trace(self):
         """
@@ -245,14 +247,14 @@ class TestActionExecutionController(FunctionalTest):
         self.assertEqual(post_resp.status_int, 201)
         delete_resp = self._do_delete(self._get_actionexecution_id(post_resp))
         self.assertEqual(delete_resp.status_int, 200)
-        self.assertEqual(delete_resp.json['status'], 'canceled')
+        self.assertEqual(delete_resp.json['status'], 'canceling')
         trace_id = str(Trace.get_all()[0].id)
         LIVE_ACTION_TRACE['context'] = {'trace_context': {'id_': trace_id}}
         post_resp = self._do_post(LIVE_ACTION_TRACE)
         self.assertEqual(post_resp.status_int, 201)
         delete_resp = self._do_delete(self._get_actionexecution_id(post_resp))
         self.assertEqual(delete_resp.status_int, 200)
-        self.assertEqual(delete_resp.json['status'], 'canceled')
+        self.assertEqual(delete_resp.json['status'], 'canceling')
 
     def test_post_parameter_validation_failed(self):
         execution = copy.deepcopy(LIVE_ACTION_1)